pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "base58>=2.1.1",
    "discord-py>=2.5.0",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.28.1",
    "solana>=0.36.5",
]
